+++ src/Rib.hs
@@ -2,8 +2,7 @@
 module Rib
   ( module Rib.App,
     module Rib.Shake,
-    module Rib.Target,
-    SourceReader,
+    module Rib.Route,
     MMark,
     Pandoc,
   )
@@ -12,6 +11,5 @@ where
 import Rib.App
 import Rib.Parser.MMark (MMark)
 import Rib.Parser.Pandoc (Pandoc)
+import Rib.Route
 import Rib.Shake
-import Rib.Source
-import Rib.Target