+++ CHANGELOG.md
@@ -14,10 +14,11 @@
   - `ghcSyntaxHighlighter`: we already have `skylighting` (which supports more parsers than Haskell)
   - `obfuscateEmail`: requires JS, which is not documented.
 - API changes:
-  - Removed `buildHtmlMulti`, `buildHtml` and `Source` type.
-  - Introduced `Target` type, and `loadTarget`/`writeTarget` functions.
-  - Introduced `forEvery` to run a Shake action over a pattern of files when they change.
+  - Introduced `Route` functionality for simpler management of static routes.
+    - Removed `buildHtmlMulti`, `buildHtml`, `readSource` functions and `Source` type.
+  - Introduced `Rib.Shake.forEvery` to run a Shake action over a pattern of files when they change.
   - Exposed `Rib.Shake.writeFileCached`
+  - `MMark.parse` and `Pandoc.parse` now automatically append path to `ribInputDir` and do not return Either.
   - Added `MMark.parseWith` (and `parsePureWith`), to specify a custom list of mmark extensions
 - Bug fixes
   - #95: Fix Shake error `resource busy (file is locked)`